+++ src/cosmos/engine/CosmosEngine.ts
@@ -63,11 +63,12 @@ import { PluginEnvironment } from '../../common/innerPlugin'
 import { getRandomDelayMs } from '../../common/network'
 import { asMaybeContractLocation } from '../../common/tokenHelpers'
 import { MakeTxParams } from '../../common/types'
-import { cleanTxLogs } from '../../common/utils'
+import { cleanTxLogs, getFetchCors } from '../../common/utils'
 import { CosmosTools } from '../CosmosTools'
 import {
   asCosmosPrivateKeys,
   asCosmosTxOtherParams,
+  asCosmosUserSettings,
   asCosmosWalletOtherData,
   asCosmosWcGetAccountsRpcPayload,
   asCosmosWcSignAminoRpcPayload,
@@ -78,6 +79,7 @@ import {
   CosmosNetworkInfo,
   CosmosOtherMethods,
   CosmosTxOtherParams,
+  CosmosUserSettings,
   CosmosWalletOtherData,
   CosmosWcRpcPayload,
   IbcChannel,
@@ -152,6 +154,7 @@ export class CosmosEngine extends CurrencyEngine<
   stakedBalanceCache: string
   stakingSupported: boolean
   chainId: string
+  declare currentSettings: CosmosUserSettings
 
   constructor(
     env: PluginEnvironment<CosmosNetworkInfo>,
@@ -161,7 +164,10 @@ export class CosmosEngine extends CurrencyEngine<
   ) {
     super(env, tools, walletInfo, opts)
     this.networkInfo = env.networkInfo
-    this.fetchCors = env.io.fetchCors
+    this.fetchCors = getFetchCors(env.io, () => {
+      const networkPrivacy = this.currentSettings?.networkPrivacy
+      return networkPrivacy === 'nym' ? { privacy: 'nym' } : {}
+    })
     this.accountNumber = 0
     this.sequence = 0
     this.unconfirmedTransactionCache = {
@@ -581,6 +587,12 @@ export class CosmosEngine extends CurrencyEngine<
 
   setOtherData(raw: any): void {
     this.otherData = asCosmosWalletOtherData(raw)
+  }
+
+  async changeUserSettings(userSettings: object): Promise<void> {
+    // Validate the user settings with our cleaner
+    asCosmosUserSettings(userSettings)
+    await super.changeUserSettings(userSettings)
   }
 
   getClients(): CosmosClients {
@@ -1019,7 +1031,7 @@ export class CosmosEngine extends CurrencyEngine<
   // // ****************************************************************************
 
   async startEngine(): Promise<void> {
-    await this.tools.connectClient()
+    await this.tools.connectClient(this.fetchCors)
     this.addToLoop('queryBalance', ACCOUNT_POLL_MILLISECONDS)
     this.addToLoop('queryBlockheight', ACCOUNT_POLL_MILLISECONDS)
     this.addToLoop('queryTransactions', TRANSACTION_POLL_MILLISECONDS)