+++ src/cosmos/info/axelarInfo.ts
@@ -7,6 +7,7 @@ import {
   CosmosInfoPayload,
   CosmosNetworkInfo
 } from '../cosmosTypes'
+import { makeCosmosDefaultSettings } from './cosmosCommonInfo'
 
 const networkInfo: CosmosNetworkInfo = {
   bech32AddressPrefix: 'axelar',
@@ -58,6 +59,7 @@ const currencyInfo: EdgeCurrencyInfo = {
   memoOptions: [{ type: 'text', maxLength: 250 }],
 
   // Deprecated:
+  defaultSettings: makeCosmosDefaultSettings(),
   displayName: 'Axelar'
 }